refactor Labeler configuration
